Move extensions for `Node` to `NodeExtensions.kt`
