+++ compiler/src/main/kotlin/net/ormr/jukkas/ast/Node.kt
@@ -18,17 +18,10 @@ package net.ormr.jukkas.ast
 
 import net.ormr.jukkas.Position
 import net.ormr.jukkas.Positionable
-import net.ormr.jukkas.reporter.MessageType
 import kotlin.properties.ReadWriteProperty
 import kotlin.reflect.KProperty
 
-val Node.closestTable: Table
-    get() = ancestors
-        .filterIsInstance<TableContainer>()
-        .firstOrNull()
-        ?.table ?: error("Could not find a table anywhere in the scope for $this")
-
-abstract class Node : Positionable {
+sealed class Node : Positionable {
     open var parent: Node? = null
 
     abstract val position: Position?
@@ -145,34 +138,4 @@ abstract class Node : Positionable {
             child = value?.let(thisRef::adopt)
         }
     }
-}
-
-/**
- * Walks up the hierarchy of `this` [Node] invoking [action] on its [parent][Node.parent] until none is left.
- *
- * Note that the first invocation of [action] will be with `this` as its argument, this differs from [Node.ancestors]
- * which *only* contains the parents of the node.
- *
- * @see [Node.ancestors]
- */
-inline fun Node.walkHierarchy(action: (Node) -> Unit) {
-    var current: Node? = this
-    while (current != null) {
-        action(current)
-        current = current.parent
-    }
-}
-
-fun Node.reportSemanticError(position: Positionable, message: String) {
-    val unit = compilationUnit
-    unit.reporter.reportError(unit.source, MessageType.Error.SEMANTIC, position, message)
-}
-
-fun Node.reportTypeError(position: Positionable, message: String) {
-    val unit = compilationUnit
-    unit.reporter.reportError(unit.source, MessageType.Error.TYPE, position, message)
-}
-
-internal fun <T : Node> T.disownParent(): T = apply {
-    parent?.disown(this)
 }
